ara: svc: es2: add tables of M-PHY fixups
These values will be used to improve link quality on ES2 silicon.

Ideally, they would in a directory that is commonly available to both
SVC and switch, but I'm not aware of a place for that, so these files
will be copied into an ES2 bridge directory in a later patch.

Testing Done: build tested; fixes themselves tested in later patches
Bug Number: SW-744
Review URL:
Reviewed by:

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>