dwc: add quiet option for dwc2 usb driver
Memory is a precious ressource and we are already close to limits on TSB. Most
of the time the DWC2 printf are useless or non informative except when
debugging errors. This patch will disable all the dwc2 printf, in order to
free the space used by the code and the strings.

Signed-off-by: Fabien Parent <fparent@baylibre.com>