greybus: unipro: don't loop unti cport is connected
Before the control protocol the landed, we needed to wait that the cport
was connected before initializing it. With the control protocol in place, we
now know that when gb_unipro_listen is called, the cport has been connected.

Don't loop until cport is connected since this function is supposed to be when
the cport has been connected.