tsb: unipro: add unipro_cport_status api
Add an API that will allow external code (its first user will be the AP), to
probe the status of a CPort. For now the only status exposed is whether the
CPort is currently in a reset state or not.

Testing Done: used that api from the kernel
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by: Fabien Parent <fparent@baylibre.com>