tsb: pinshare: remove ES2 ifdefs
This commit is preparing the ES3 support. Since we don't support ES1 anymore
and that the pinshare setting is bit compatible between ES2 and ES3 we can
get rid of the ifdef specific to ES2.

This commit also sort the pinshare macro in a ascending order.

Testing Done: gpio demo
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by: Fabien Parent <fparent@baylibre.com>