tsb: remove NB_GPIO_IRQS
On TSB bridges the number of GPIOs depends on whether it is a APBridge or
GPBridge that is running. So we can't export a static number, so lets call
tsb_gpio_line_count() via tsb_nr_gpio() in order to get the GPIO line count.

This commit is a preparation for another commit that will fix the GPIO line
count on APBridges.

Testing Done: gpio demo
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by: Fabien Parent <fparent@baylibre.com>