tsb: unipro: fix memory leak and use-after-free bug
If the unipro_reset_cport is called twice in a row and that the first reset
hasn't been completed yet, we will overwrite the private data of the first
call. Return -EINPROGRESS if a CPort reset is already in progress.

Testing Done: gpio demo. checked that reusing CPort is still successful.
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by: Fabien Parent <fparent@baylibre.com>